The pedestrian Traffic Light by tiva
The traffic light shall stay GREEN for 5 seconds, then YELLOW for 2 
seconds, then turns RED for 3 seconds.
Use one push button and 2 LEDs for this system. 
There will be two pedestrian traffics, each with a push button and 2 LEDS: 
GREEN and RED. 
Whenever a pedestrian presses the button, the traffic light that is GREEN shall 
be interrupted and changes to RED, and the pedestrian traffic light will be 
green for 2 seconds. Then it will be back to RED and the traffic light that was 
interrupted will resume from when it was paused.
Example: If the car traffic light was green for 2 seconds and interrupted it will 
then resume the remaining 3 seconds to finish the 5 seconds